LayoutTests/http/wpt/webcodecs/webcodecs-message-gc-expected.txt

CONSOLE MESSAGE: A VideoFrame was destroyed without having been closed explicitly

PASS VP8 codec GC without calling VideoFrame.close
